requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
langchain_community==0.3.16
langchain_core==0.3.33
langchain_google_genai==2.0.9
langgraph==0.2.68
protobuf==5.29.3
python-dotenv==1.0.1
gunicorn==23.0.0